challenges/plxs/requirements.txt
attrs==21.2.0
Cerberus==1.3.4
click==8.0.1
coverage==5.5
Flask==2.0.1
Flask-Cors==3.0.10
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
packaging==21.0
pluggy==1.0.0
py==1.10.0
pyparsing==2.4.7
pytest==6.2.5
pytest-cov==2.12.1
python-dotenv==0.19.0
six==1.16.0
toml==0.10.2
Werkzeug==2.0.1
